+++ include/solver.h
@@ -21,20 +21,20 @@ protected:
 
 class BaselineSolver final : public Solver {
 public:
-    BaselineSolver(double st, int aon, int max_iter, int time_limit, int verbose = 2);
+    BaselineSolver(double st, int aon, int max_iter, double time_limit, int verbose = 2);
     ~BaselineSolver() override = default;
 };
 
 class AnnealingSolver final : public Solver {
 public:
-    AnnealingSolver(double st, int aon, int max_iter, int time_limit, double alpha,
+    AnnealingSolver(double st, int aon, int max_iter, double time_limit, double alpha,
                     int verbose = 2);
     ~AnnealingSolver() override = default;
 };
 
 class RebalancingSolver final : public Solver {
 public:
-    RebalancingSolver(double st, int aon, int max_iter, int time_limit, double fairness,
+    RebalancingSolver(double st, int aon, int max_iter, double time_limit, double fairness,
                       int verbose = 2);
     ~RebalancingSolver() override = default;
 };
